Move merge functions with their types
