+++ report/topology.go
@@ -5,8 +5,6 @@ import (
 	"strings"
 )
 
-const localUnknown = "localUnknown"
-
 // Topology describes a specific view of a network. It consists of nodes and
 // edges, represented by Adjacency, and metadata about those nodes and edges,
 // represented by EdgeMetadatas and NodeMetadatas respectively.
@@ -16,17 +14,51 @@ type Topology struct {
 	NodeMetadatas
 }
 
+// Merge merges another Topology into the receiver.
+func (t *Topology) Merge(other Topology) {
+	t.Adjacency.Merge(other.Adjacency)
+	t.EdgeMetadatas.Merge(other.EdgeMetadatas)
+	t.NodeMetadatas.Merge(other.NodeMetadatas)
+}
+
 // Adjacency is an adjacency-list encoding of the topology. Keys are node IDs,
 // as produced by the relevant MappingFunc for the topology.
 type Adjacency map[string]IDList
+
+// Merge merges another Adjacency list into the receiver.
+func (a *Adjacency) Merge(other Adjacency) {
+	for addr, adj := range other {
+		(*a)[addr] = (*a)[addr].Merge(adj)
+	}
+}
 
 // EdgeMetadatas collect metadata about each edge in a topology. Keys are a
 // concatenation of node IDs.
 type EdgeMetadatas map[string]EdgeMetadata
 
+// Merge merges another EdgeMetadatas into the receiver. If other is from
+// another probe this is the union of both metadatas. Keys present in both are
+// summed.
+func (e *EdgeMetadatas) Merge(other EdgeMetadatas) {
+	for id, edgemeta := range other {
+		local := (*e)[id]
+		local.Merge(edgemeta)
+		(*e)[id] = local
+	}
+}
+
 // NodeMetadatas collect metadata about each node in a topology. Keys are node
 // IDs.
 type NodeMetadatas map[string]NodeMetadata
+
+// Merge merges another NodeMetadatas into the receiver.
+func (m *NodeMetadatas) Merge(other NodeMetadatas) {
+	for id, meta := range other {
+		if _, ok := (*m)[id]; !ok {
+			(*m)[id] = meta // not a copy
+		}
+	}
+}
 
 // EdgeMetadata describes a superset of the metadata that probes can possibly
 // collect about a directed edge between two nodes in any topology.
@@ -36,6 +68,28 @@ type EdgeMetadata struct {
 	EgressByteCount    *uint64 `json:"egress_byte_count,omitempty"`  // Transport layer
 	IngressByteCount   *uint64 `json:"ingress_byte_count,omitempty"` // Transport layer
 	MaxConnCountTCP    *uint64 `json:"max_conn_count_tcp,omitempty"`
+}
+
+// Merge merges another EdgeMetadata into the receiver. The two edge metadatas
+// should represent the same edge on different times.
+func (m *EdgeMetadata) Merge(other EdgeMetadata) {
+	m.EgressPacketCount = merge(m.EgressPacketCount, other.EgressPacketCount, sum)
+	m.IngressPacketCount = merge(m.IngressPacketCount, other.IngressPacketCount, sum)
+	m.EgressByteCount = merge(m.EgressByteCount, other.EgressByteCount, sum)
+	m.IngressByteCount = merge(m.IngressByteCount, other.IngressByteCount, sum)
+	m.MaxConnCountTCP = merge(m.MaxConnCountTCP, other.MaxConnCountTCP, max)
+}
+
+// Flatten sums two EdgeMetadatas. Their windows should be the same duration;
+// they should represent different edges at the same time.
+func (m *EdgeMetadata) Flatten(other EdgeMetadata) {
+	m.EgressPacketCount = merge(m.EgressPacketCount, other.EgressPacketCount, sum)
+	m.IngressPacketCount = merge(m.IngressPacketCount, other.IngressPacketCount, sum)
+	m.EgressByteCount = merge(m.EgressByteCount, other.EgressByteCount, sum)
+	m.IngressByteCount = merge(m.IngressByteCount, other.IngressByteCount, sum)
+	// Note that summing of two maximums doesn't always give us the true
+	// maximum. But it's a best effort.
+	m.MaxConnCountTCP = merge(m.MaxConnCountTCP, other.MaxConnCountTCP, sum)
 }
 
 // NodeMetadata describes a superset of the metadata that probes can collect
@@ -56,6 +110,20 @@ func MakeNodeMetadataWith(m map[string]string) NodeMetadata {
 		Metadata: m,
 		Counters: map[string]int{},
 	}
+}
+
+// Merge merges two node metadata maps together. In case of conflict, the
+// other (right-hand) side wins. Always reassign the result of merge to the
+// destination. Merge is defined on the value-type, but node metadata map is
+// itself a reference type, so if you want to maintain immutability, use copy.
+func (nm NodeMetadata) Merge(other NodeMetadata) NodeMetadata {
+	for k, v := range other.Metadata {
+		nm.Metadata[k] = v // other takes precedence
+	}
+	for k, v := range other.Counters {
+		nm.Counters[k] = nm.Counters[k] + v
+	}
+	return nm
 }
 
 // Copy returns a value copy, useful for tests.
@@ -139,4 +207,26 @@ func (t Topology) Validate() error {
 	}
 
 	return nil
+}
+
+func merge(dst, src *uint64, op func(uint64, uint64) uint64) *uint64 {
+	if src == nil {
+		return dst
+	}
+	if dst == nil {
+		dst = new(uint64)
+	}
+	(*dst) = op(*dst, *src)
+	return dst
+}
+
+func sum(dst, src uint64) uint64 {
+	return dst + src
+}
+
+func max(dst, src uint64) uint64 {
+	if dst > src {
+		return dst
+	}
+	return src
 }
